STCOM-934: fixed comments
