+++ hooks/useAdvancedSearch/index.js
@@ -1,0 +1,2 @@
+
+export { default } from './useAdvancedSearch';
