+++ lib/AdvancedSearch/AdvancedSearch.js
@@ -1,36 +1,15 @@
-import {
-  useCallback,
-  useState,
-  useMemo,
-  useEffect,
-} from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 
+import { ROW_COUNT } from './constants';
 import AdvancedSearchModal from './components/AdvancedSearchModal';
 import AdvancedSearchRow from './components/AdvancedSearchRow';
+import useAdvancedSearch from '../../hooks/useAdvancedSearch';
 import defaultQueryBuilder from './utilities/defaultQueryBuilder';
-import {
-  ROW_COUNT,
-  FIELD_NAMES,
-  QUERY_OPTION_VALUE,
-  BOOLEAN_OPERATORS,
-} from './constants';
+import filterFilledRows from './utilities/filterFilledRows';
 
 const defaultRowFormatter = (searchOption, query, comparator) => {
   return `${searchOption}${comparator}${query}`;
-};
-
-const createInitialRowState = (firstRowInitialSearch, defaultSearchOptionValue) => {
-  return new Array(ROW_COUNT).fill(null).map((_, index) => ({
-    [FIELD_NAMES.BOOL]: BOOLEAN_OPERATORS.AND,
-    [FIELD_NAMES.QUERY]: index === 0
-      ? firstRowInitialSearch?.query || ''
-      : '',
-    [FIELD_NAMES.SEARCH_OPTION]: index === 0
-      ? firstRowInitialSearch?.option || defaultSearchOptionValue
-      : defaultSearchOptionValue,
-  }));
 };
 
 const propTypes = {
@@ -64,106 +43,26 @@ const AdvancedSearch = ({
   children,
 }) => {
   const intl = useIntl();
-
-  const initialRowState = useMemo(() => {
-    return createInitialRowState(firstRowInitialSearch, defaultSearchOptionValue);
-  }, [firstRowInitialSearch, defaultSearchOptionValue]);
-
-  const [rowState, setRowState] = useState(initialRowState);
-
-  const searchOptionsWithQuery = useMemo(() => (
-    [{
-      label: intl.formatMessage({ id: 'stripes-components.advancedSearch.searchOptions.query' }),
-      value: QUERY_OPTION_VALUE,
-    }, ...searchOptions]
-  ), [intl, searchOptions]);
-
-  const [showEmptyFirstRowMessage, setShowEmptyFirstRowMessage] = useState(false);
-
-  const resetRows = () => {
-    setRowState(createInitialRowState(firstRowInitialSearch, defaultSearchOptionValue));
-  };
-
-  const filterFilledRows = (rows) => {
-    return rows.filter(row => !!row[FIELD_NAMES.QUERY]);
-  };
-
-  const onChange = (rowIndex, key, value) => {
-    setRowState((currentRowState) => {
-      const newRowState = [...currentRowState];
-
-      newRowState[rowIndex][key] = value;
-
-      return newRowState;
-    });
-  };
-
-  const splitQueryRows = (rows) => {
-    const splitRow = (row) => {
-      const queryString = row[FIELD_NAMES.QUERY];
-
-      const splitIntoRowsRegex = /((?:or|and|not)\s+)?([\w]+==[^\s]+)/g;
-
-      const matches = [...queryString.matchAll(splitIntoRowsRegex)];
-
-      return matches.map((match) => {
-        const [, operator = '', query] = match;
-
-        const [option, value] = query.split('==').map(i => i.trim());
-
-        return {
-          [FIELD_NAMES.QUERY]: value,
-          [FIELD_NAMES.BOOL]: operator.trim(),
-          [FIELD_NAMES.SEARCH_OPTION]: option.trim(),
-        };
-      });
-    };
-
-    return rows.reduce((finalRows, row) => {
-      if (row[FIELD_NAMES.SEARCH_OPTION] === QUERY_OPTION_VALUE) {
-        // split row with multiple search options into array of rows
-
-        const splittedRow = splitRow(row);
-
-        return [...finalRows, ...(splittedRow.length ? splittedRow : [{ bool:'and', search: '', searchOption: 'keyword' }])];
-      }
-
-      return [...finalRows, row];
-    }, []);
-  };
-
-  const handleSearch = useCallback(() => {
-    const splitRows = filterFilledRows(splitQueryRows(rowState));
-
-    if (!splitRows.length) {
-      setShowEmptyFirstRowMessage(true);
-
-      return;
-    }
-
-    setShowEmptyFirstRowMessage(false);
-
-    const query = queryBuilder(splitRows, rowFormatter);
-
-    onSearch(query, splitRows);
-  }, [queryBuilder, onSearch, rowState, rowFormatter]);
-
-  const handleInitialSearchChange = () => {
-    const splitRows = splitQueryRows(initialRowState);
-
-    setRowState(splitRows);
-  };
-
-  useEffect(() => {
-    handleInitialSearchChange();
-  }, [firstRowInitialSearch]);
-
-  const handleCancel = useCallback(() => {
-    onCancel();
-  }, [onCancel]);
+  const {
+    handleCancel,
+    handleSearch,
+    onChange,
+    resetRows,
+    rowState,
+    searchOptionsWithQuery,
+    showEmptyFirstRowMessage,
+  } = useAdvancedSearch({
+    defaultSearchOptionValue,
+    firstRowInitialSearch,
+    onCancel,
+    onSearch,
+    queryBuilder,
+    rowFormatter,
+    searchOptions,
+  });
 
   const renderRows = () => {
-    return new Array(rowState.length).fill(null).map((_, index) => (
+    return new Array(ROW_COUNT).fill(null).map((_, index) => (
       <AdvancedSearchRow
         key={`$advanced-search-row-${index}`}
         index={index}
@@ -178,8 +77,6 @@ const AdvancedSearch = ({
     ));
   };
 
-  const filledRows = filterFilledRows(rowState);
-
   return (
     <>
       <AdvancedSearchModal
@@ -190,7 +87,7 @@ const AdvancedSearch = ({
       >
         {renderRows()}
       </AdvancedSearchModal>
-      {children ? children({ resetRows, rowState: filledRows }) : null}
+      { children ? children({ resetRows, rowState: filterFilledRows(rowState) }) : null }
     </>
   );
 };
